+++ src/MainDashBoard/components/mainDashBoardOkrTree/MainDashboardOKRTree.tsx
@@ -36,12 +36,13 @@ const MainDashboardOKRTree = ({ onShowSideSheet }: IMainDashboardOKRTreeProps) =
 export default MainDashboardOKRTree;
 
 const okrTreeContainer = css`
-  position: relative;
   display: flex;
   align-items: center;
   justify-content: center;
   width: 100%;
-  height: 100%;
+  height: calc(100% - 7.6rem);
+  margin-right: 23.2rem;
+  margin-bottom: 7.6rem;
 `;
 
 const okrTree = css`
